+++ plugins/sonarqube-actions/src/alpha.ts
@@ -1,0 +1,1 @@
+export { scaffolderModuleSonarqubeActions as default } from './module';
